% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar plots}
\usage{
bar_plot(height, type = 1, ...)
}
\arguments{
\item{height}{The argument \code{height} of \code{\link[graphics]{barplot}}}

\item{type}{An integer between 1 and 4 determining the style of the plot.
\itemize{
\item 1 : default from \code{barplot}
\item 2 : default barplot plus thin baseline
\item 3 : default barplot plus thin baseline
\item 4 : default barplot in R
}}
}
\value{
See the \strong{Value} section of \code{\link[graphics]{barplot}}.
}
\description{
Uses \code{\link[graphics]{barplot}} to produce different forms of barplot
based on Tufte. E. (2001) The visual display of quantitative information.
}
\details{
See the \strong{Details} section of \code{\link[graphics]{barplot}}.
}
