% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process_movie.R
\name{poisson_process_movie}
\alias{poisson_process_movie}
\title{Poisson process movie}
\usage{
poisson_process_movie(lambda = 1, hours = 24, pos = 1,
  envir = as.environment(pos))
}
\arguments{
\item{lambda}{A numeric scalar.  The rate parameter of the exponential
distribution from which data are to be simulated using \code{\link{rexp}}.}

\item{hours}{A positive integer scalar.  The number of hours for which to
simulate a Poisson process of rate \eqn{lambda} events per hour.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate that if events arrive in a Poisson process of
rate \eqn{\lambda} per hour then the number of events that arrive
during a fixed interval of length \eqn{t} hours has a Poisson
distribution with mean \eqn{\lambda t}.
}
\details{
This movie consists of two plots: one on the top of the display
  the other on the bottom.

  The top plot displays the times at which events simulated from a
  Poisson process of rate \eqn{\lambda} events per hour occur during
  the time intervals (0, \code{hours} hours.  The total numbers of events
  that occur in each hour are also displayed on the plot.  Each time the
  button "simulate another sequence of events" is clicked a new set of
  simulated events is produced.

  The bottom plot contains a barplot (the black bars) summarising how
  often the an hour contains 0, 1, 2, ... events.  Specifically, the
  proportions of the hours for which there are 0, 1 ,2, ... events is
  plotted on the vertical axis.  This is an estimate of the probability
  mass function of the random variable defined as the number of events
  that occur in a one hour period, when those event occur in a Poisson
  process of rate \eqn{\lambda} per hour.  Each time a set of events
  is simulated the totals in each hour are \strong{added} to the
  current collection of totals.

  Also included in the bottom plot is a barplot (the red bars) showing
  the p.m.f. of a Poisson(\eqn{\lambda}) random variable.  As the number
  of totals added to plot increasing the black bars become closer and
  closer to the red bars.

  Note also that the independence assumption underlying the Poisson
  process means that the numbers of events occurring in different
  hours are independent.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

# Produce movie using values from the Aussie births data
\dontrun{
poisson_process_movie(lambda = 2, hours = 24)
}
}
\seealso{
\code{\link{movies}}: general information about STAT1004 movies.
}
