% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot_movie.R
\name{qq_plot_movie}
\alias{qq_plot_movie}
\title{Quantile-Quantile plot movie}
\usage{
qq_plot_movie(data = NULL, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{A numeric vector.  Data on which to base the QQ plot.}

\item{mu, sigma}{Numeric scalars.  The mean and standard deviation of
the normal distribution to which the data \code{data} are to be compared.
If \code{mu} and/or \code{sigma} are not supplied then they are estimated
from the data.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate the constructions of a normal Quantile-Quantile
(QQ) plot.
}
\details{
Add details.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

# Data in which an event occurs exactly on the hour
\dontrun{
# Create the same dataset used in the lecture notes/slides
set.seed(382)
x <- sort(round(rnorm(9, mean = 7.72, sd = 3.57), 2))
qq_plot_movie(data = x, mu = 7.72, sigma = 3.57)
}
}
\seealso{
\code{\link{movies}}: general information about STAT1004 movies.
}
