% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_poisson_process.R
\name{simulate_poisson_process}
\alias{simulate_poisson_process}
\title{Simulation of a Poisson process}
\usage{
simulate_poisson_process(lambda = 1, hours = 24, n_events = NULL)
}
\arguments{
\item{lambda}{A positive numeric scalar.  The rate of the Poisson process.}

\item{hours}{A positive integer scalar.  The number of hours for which to
simulate a Poisson process of rate \eqn{lambda} events per hour.}

\item{n_events}{A positive integer scalar.  The number of events to
simulate.}
}
\value{
A numeric vector containin the times at which the events occur.
}
\description{
Simulates event times from a (one-dimensional, homogeneous) Poisson process
of rate \eqn{\lambda} per hour.  The user has the options to simulate events
over a fixed time period of \code{hours} hours or to simulate a fixed number
\code{n_events} of events.
}
\details{
If \code{n_events} is supplied then exactly \code{n_events} are
  simulated and \code{hours} has no effect.  If \code{n_events} is not
  supplied then events are simulated over the time interval
  (0, \code{hours}).  If no events occur in (0, \code{hours}) then
  the value -1 is returned.
}
