% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{skew}
\alias{skew}
\title{Standardized sample skewness}
\usage{
skew(x, type = 6, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{na.rm}{A logical scalar.  If true, any \code{\link{NA}} and NaN's
are removed from \code{x} before the quantiles are computed.}
}
\value{
A numeric scalar (if the input was a vector) or vector (if the input
  was a matrix).
}
\description{
Calculates the standardized sample skewness of a vector of data or of each
column of a matrix of data, based on the estimators of the lower quartile,
median and upper quartile in the STAT1004 notes.
}
\details{
See Section 2.3 of the
\href{https://moodle.ucl.ac.uk/pluginfile.php/3054127/mod_resource/content/22/STAT1004notes-2x1.pdf}{STAT1004 notes.}
}
\examples{
birth_times <- ox_births[, "time"]
skew(birth_times)
}
\seealso{
\code{\link[stats]{quantile}} for calculating sample quantiles.
}
