% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat1004.R
\docType{data}
\name{world_bank}
\alias{world_bank}
\alias{co2_pc}
\alias{gdp_pc}
\alias{popn}
\alias{life}
\title{World Bank Data}
\format{A data frame with 217 rows (one for each country) and one column
  for each year in the data, e.g. 55 columns for \code{co2_pc}.
  The names of the countries are given in the row names of the dataset,
  e.g. by \code{rownames(popn)}.}
\source{
The World Bank's
\href{https://data.worldbank.org/data-catalog/world-development-indicators}{World Development Indicators database}.
}
\usage{
co2_pc

gdp_pc

popn

life
}
\description{
Selected annual World Development Indicators provided in the World Bank's
World Development Indicators database.  The indicators are
\itemize{
  \item{\strong{co2_pc}: }{total carbon dioxide emissions per capita, in metric tons
    per person, 1960-2014}
  \item{\strong{gdp_pc}: }{Gross Domestic Product (GDP) per capita, in current
    US dollars per person, 1960-2016}
  \item{\strong{popn}: }{total population size, 1960-2016}
  \item{\strong{life}: }{life expectancy at birth, in years, 1960-2015}
}
}
\keyword{datasets}
