% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{movies}
\alias{movies}
\alias{shuttle_movie}
\title{Movies}
\usage{
shuttle_movie(n_reps = 1, pos = 1, envir = as.environment(pos))
}
\arguments{
\item{n_reps}{An integer scalar.  Relevant to \code{shuttle_movie} only.
The number of flights to simulate
for each of the 23 (pre-disaster) temperatures in the real dataset.
For example, \code{n_reps = 10} means that we simulate a dataset of
size 230.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
Animations used to illustrate key ideas in STAT1004.  The animations
are produced using the package \code{\link[rpanel]{rpanel}}.
The rpanel package must be loaded, using
`library(rpanel)`, in order that these functions work.
If \code{rpanel} is not loaded then an error message similar to
\code{Error in shuttle_movie() : could not find function "rp.control"}
will be produced.
}
\details{
When one of thes functions is called R opens up a small
\emph{parameter window} containing clickable buttons that can be
used to change parameters underlying the plot. For the effects of
these buttons see the documentation of the individual functions below.

The parameter window does not close automatically after the movie:
the user needs to close it manually.

Some movies create objects in the global environment, that is, objects
that will be listed when \code{ls()} is used.  `rm` can be used to remove
these objects if desired.  For example `rm(name)` can be used to remove
object `name`.
}
\section{Space shuttle movie}{

A movie to illustrate uncertainty in the linear logistic regression curves
fitted to the real space shuttle data. \code{shuttle_movie} produces an
animated version of the plot produced by \code{\link{shuttle_sim_plot}}.
Calling `shuttle_movie` creates a plot containing the real data
and the logistic curve fitted to the real data and opens a
parameter window containing buttons labelled + and -.
Clicking + adds to the plot a new simulated dataset and the logistic curve
fitted to this dataset.
Clicking - removes the mostly recently added dataset.
}

\examples{
# Load package rpanel
library(rpanel)

# Challenger space shuttle data -------------

# Movie based on datasets of size 23
shuttle_movie()
# Movie based on datasets of size 230
shuttle_movie(n_reps = 10)

}
\references{
Bowman, A., Crawford, E., Alexander, G. and Bowman, R. W.
 (2007). rpanel: Simple Interactive Controls for R Functions Using the
 tcltk Package.  \emph{Journal of Statistical Software}, \strong{17(9)}, 1-18.
 \url{http://www.jstatsoft.org/v17/i09/}.
}
